+++ src/backend/core/external_api/viewsets.py
@@ -61,7 +61,7 @@ class ApplicationViewSet(viewsets.ViewSet):
         except models.Application.DoesNotExist as e:
             raise drf_exceptions.AuthenticationFailed("Invalid credentials") from e
 
-        if not application.active:
+        if not application.is_active:
             raise drf_exceptions.AuthenticationFailed("Application is inactive")
 
         if not check_password(client_secret, application.client_secret):